Use 32 bit floats in Map module to use less memory on GPU

+++ src/map/c2_map.jl
@@ -28,7 +28,7 @@ function Params_C2(img; periodic::Bool=true)
         periodic,
         total,
         similar(img, Int),
-        similar(img, ComplexF64, complex_box),
+        similar(img, ComplexF32, complex_box),
     )
     cf_type = periodic ? :periodic_point_point : :central_symmetry
     p, cf_type
@@ -70,7 +70,7 @@ The `image` contains the probability of the voxel being in the correct phase.
 # Examples
 ```jldoctest
 julia> c2([1 0; 0 1]; periodic=true).result
-2×2 Matrix{Float64}:
+2×2 Matrix{Float32}:
  0.5  0.0
  0.0  0.0
 ```
